requirements.txt
csv
imagehash
Pillow
setuptools
time